requirements.txt
numpy==1.25.2
packaging==23.1
Pillow==10.0.0
pytesseract==0.3.10
